fix structure and style
