+++ portalCheckIn/server.js
@@ -8,6 +8,7 @@ var index         = require('./server/route/index');
 var users         = require('./server/route/usersRoute');
 var admins        = require('./server/route/admin_routes');
 var mongoose      = require('mongoose');
+var get_schedule = require('./get_schedule');
 
 
 //connect to mongo
