+++ portalCheckIn/server/model/actionModel.js
@@ -3,20 +3,29 @@
  */
 var mongoose = require('mongoose');
 
+var commentSchema = mongoose.Schema({
+    body: String,
+    user: {
+        _id: Number,
+        name: String
+    },
+    createdAt: {type: Date, default: Date.now}
+});
+
 actionSchema = mongoose.Schema({
     type: [String],
     //whether we need id?
     user: {
-        studentId:Number,
+        _id:Number,
         name:String
     },
 
-    createdAt: Date,
-    comments: Array,
+    createdAt: {type: Date, default: Date.now},
+    comments: [commentSchema],
     extras: Object
 });
 
-
+actionSchema.index({ type: 1, 'user._id': 1, createdAt: -1 });
 
 var action = mongoose.model("action", actionSchema);
 module.exports = action;
